Made it so percentComplete on dreams is based on tasks percentages

+++ Dream_Manager/grails-app/services/dream_manager/TaskService.groovy
@@ -6,7 +6,7 @@ class TaskService {
 
     }
 	
-	def getMaxOrder(int dream){
+	def getMaxOrder(long dream){
 		def taskList = Task.withCriteria{eq('dream', Dream.findById(dream))}
 		taskList.size()+1
 	}
